Almost Final Update

+++ BillingProject/src/main/java/com/jkt/reimbursement/service/BillService.java
@@ -4,7 +4,6 @@ import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -25,7 +24,17 @@ public class BillService {
 		billRepo.save(bill);
 	}
 	
-
+	public void updateBill(Bill bill,int id)
+	{
+		Bill b1=billRepo.findById(id);
+		if(!b1.equals(null))
+		{
+			b1.setRemark(bill.getRemark());
+			billRepo.save(b1);
+		}
+		else
+			throw new RuntimeException("Bill not found-" +id);
+	}
 	//mapping add bill user/bill
 	public void AddBill(Bill bill,MultipartFile file)
 	{
@@ -48,33 +57,29 @@ public class BillService {
 	public void delBill(int id)
 	{
 		Bill bill=billRepo.getOne(id);
-		billRepo.delete(bill);
-	}
-	
-	public Bill findById(int id)
-	{
-	Optional<Bill> result=billRepo.findById(id);
-	
-	Bill thebill=null;
-	if(result.isPresent())
-	{
-		thebill=result.get();
-	}
-	return thebill;
+		if(!bill.equals(null)) {
+			billRepo.delete(bill);
+		}
+		else
+			throw new RuntimeException("Bill not found-" +id);
 	}
 	
 	//mapping find bill by username 
-		public List<Bill> getAllBillByUsersId(String username)
-		{
-			List<Bill> bill=new ArrayList<>();
-			billRepo.findByUserId(username).forEach(bill::add);
-			return bill;
-		}
-
+	public List<Bill> getAllBillByUsersId(String username)
+	{
+		List<Bill> bill=new ArrayList<>();
+		billRepo.findByUserId(username).forEach(bill::add);
+		return bill;
+	}
 		
-		//new for file
-		public Bill getFile(int id) throws FileNotFoundException {
-			return billRepo.findById(id)
-					.orElseThrow(()->new FileNotFoundException("File not found with id" +id));
+	//new for file
+	public Bill getFile(int id) throws FileNotFoundException 
+	{
+		Bill b1=billRepo.findById(id);
+		if(!b1.equals(null)) {
+			return billRepo.findById(id);
 		}
+		else
+			throw new RuntimeException("File not found-" +id);
+	}
 }
